put rescue blocks around test code that breaks in Rails 3
multi_rails appears to be incompatible with Rails 3, and some of the
acts_as_audited code (require 'action_controller/test_process') is,
too.  I just added rescue blocks to allow the tests to proceed.

FIXME I don't have a replacement for
ActionController::Routing::Routes.draw so tests that depend on that
are broken.  At the moment we've got 4 errors and one failure but at
least the tests run.